+++ CHANGES.rst
@@ -12,6 +12,8 @@ Changelog
   [sgeulette]
 - In IncomingEmail consumer, when the agent has a primary organization, we select it to set treating_groups.
   [sgeulette]
+- In IncomingEmail consumer, when the agent is part of the encodeurs group, treating_groups is not set.
+  [sgeulette]
 
 0.19 (2023-07-07)
 -----------------